[TASK] Avoid unsolveable phpstan error for return-type mismatch
PHP8.1 introduced native return types to native methods
and interfaces, which crashes if they are not compatible.
As a workaround the introduced `#[\ReturnTypeWillChange]`
attribute can be used to mitigate this, which has been
used to mitigate quite some issues with it.

PHPStan has released a new version which fixes a bunch
of bugs, which detects more issues in core. These will
be addressed with the dedicated PHPStan raise patch #98246.
However, one issue cannot be solved or added to the baseline
and must be fixed beforehand. The reason is, that phpstan
also evaluates the PHPDoc return type.

This change uses the `#[\ReturnTypeWillChange]` attribute
to silence this as a backportable intermediate solution.

Resolves: #98251
Related: #98246
Releases: main, 11.5
Change-Id: I24e7d24e735d2dffbca0eca21de602c40f7e7d2a
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/75575
Tested-by: core-ci <typo3@b13.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>