[TASK] Avoid PHP8.2 related deprecation failure in unit test
With PHP8.2 setting dynamic properties has been deprecated. This
is also valid if it is set through unserialization and emits a
deprecation warning. Thus this fails hard in tests.

Thus one test fails, which is expclicitly testing unserialization
with invalid dynamic class properties.

This change moves the test to unitDeprecated tests, which will not
fail on deprecation warnings. This acts pre-patch to update PHP8.2
testing to `PHP8.2.0 RC1`.

Resolves: #98266
Releases: main, 11.5
Change-Id: I7bf87b0ade906727f6222d31a1b735f452717138
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/75616
Tested-by: core-ci <typo3@b13.com>
Tested-by: Oliver Klee <typo3-coding@oliverklee.de>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Oliver Klee <typo3-coding@oliverklee.de>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>