[!!!][FEATURE] Add generic language overlay PSR-14 events instead of hooks
New PSR-14 events are introduced in favor of old hooks
within PageRepository which are related to overlaying
records. This is now achieved by using
PageRepository->getLanguageOverlay() consistently
which now contains the events both for pages and any other
kind of translatable record.

The main benefit of the events are that they consistently
work with LanguageAspect instead of integers.

Resolves: #98303
Releases: main
Change-Id: I6cf64b77e9df8581f8aafc3b6a41f4690d277830
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/75626
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Susanne Moog <look@susi.dev>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Susanne Moog <look@susi.dev>
Reviewed-by: Benni Mack <benni@typo3.org>