[!!!][TASK] Make the AbstractPlugin @internal
`\TYPO3\CMS\Frontend\Plugin\AbstractPlugin` is not a recommended way of
developing Frontend Plugins anymore. As long as it is neither deprecated
nor internal we cannot remove it from the documentation.
Also, it is one of the features that let TYPO3 appear outdated.

There are still several non-Extbase plugins that base on
`AbstractPlugin`. However, in most cases, the same result can be achieved
by removing `AbstractPlugin` as parent class.

In the rare case where functionality of `AbstractPlugin` is still used,
those methods can be copied into the affected plugin.

Releases: main
Resolves: #98281
Change-Id: I22e9bc2784f6bc0feaf38a4be1e7e1461f025620
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/75655
Tested-by: core-ci <typo3@b13.com>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Oliver Klee <typo3-coding@oliverklee.de>
Reviewed-by: Thomas Löffler <loeffler@spooner-web.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>