[BUGFIX] Do not render clipboard errors as notification
If a backend user switches pages in the list module right after the
clipboard is being loaded, the error message "An error occurred while
fetching clipboard data" may arise as a "global" notification which is
technically correct but confuses users.

To relax the situation, the error message is now rendered as a regular
alert box where the clipboard is expected to be and only if there is a
real error.

Resolves: #98242
Releases: main, 11.5
Change-Id: Ia1fddba64ac459601789aa7f68151840ddeadb3c
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/75661
Tested-by: Benjamin Franzke <bfr@qbus.de>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Benjamin Franzke <bfr@qbus.de>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>