[TASK] Provide more details when calling TSFE->set_no_cache()
When TSFE->set_no_cache() is called without a reason
(usually through third-party code), the Logger now
ships with more information such as the class name and
function / method name, which was previously just the
filename.

If the code was called from outside a PHP class,
which is very uncommon in 2022, then just the filename
without the path is now put into the context of the Logger.

This way, a nasty hack around resolving symlinks
can be removed.

Resolves: #98302
Releases: main
Change-Id: I4dd349162704e51cb49a463405107039401a82b3
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/75667
Tested-by: core-ci <typo3@b13.com>
Tested-by: Susanne Moog <look@susi.dev>
Tested-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Susanne Moog <look@susi.dev>
Reviewed-by: Oliver Bartsch <bo@cedev.de>