[TASK] Remove wrong doc comments in \TYPO3\CMS\Core\Database\Connection
The doc comments for some functions in `\TYPO3\CMS\Core\Database\Connection`
state, that table expressions and columns are not escaped. This is actually
not true and it seems those doc comments have been adopted directly from
`doctrine/dbal` back in 2016, when Doctrine DBAL was introduced to TYPO3.

Resolves: #98318
Releases: main, 11.5, 10.4
Change-Id: If23d568b23ef1b3c6f40efd50e907af54c349b3d
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/75650
Reviewed-by: Oliver Klee <typo3-coding@oliverklee.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>