[TASK] Add notification when Grunt "watch" finished compilation
This outputs a "bell" character to notify the developer that the "watch"
task has picked up and finished compilation of changes.

"bell" is meant to draw user attention and modern developer-oriented
terminals usually implement it by flashing the title bar or showing a
notification.
The action is usually configurable in the terminal settings.

This is basically https://www.npmjs.com/package/grunt-bell.

Resolves: #98326
Releases: main
Change-Id: Ia462fcd6ba2a24c56332b481895e29be02a1771f
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/75685
Tested-by: core-ci <typo3@b13.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>