[BUGFIX] Rename modal's title property to modalTitle
In #98288 the modal component was rewritten to be a lit element. The
modal title was defined by a `title` attribute. However, using this
attribute has the side effect that a tooltip is rendered wherever the
cursor is positioned.

Unfortunately, this behavior cannot get disabled in browsers. To
circumvent this issue, the attribute is now renamed to `modalTitle`.

Resolves: #98355
Related: #98288
Releases: main
Change-Id: I585bfb7c69376e8a38f85e48393625d16235b31c
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/75735
Tested-by: core-ci <typo3@b13.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Oliver Klee <typo3-coding@oliverklee.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Benni Mack <benni@typo3.org>