[TASK] Clean up AbstractUserAuthentication
This change renames variables and splits up
some logic in AbstractUserAuthentication in order
to allow further changes.

Various variable names are changed in order
to understand the logic in the main monster
"checkAuthentication" method better.

In addition, the ->getAuthInfoArray() method
is now moved to a separate call so it is obvious
what is needed, making some unit tests a bit nicer.

Resolves: #98385
Releases: main
Change-Id: I447b742b88520e5ec8d5c27628fea715dde40110
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/75797
Tested-by: core-ci <typo3@b13.com>
Tested-by: Oliver Hader <oliver.hader@typo3.org>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Oliver Hader <oliver.hader@typo3.org>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>