[BUGFIX] Do not override width of preview mode selector
The preview mode selector doesn't change its width anymore based on the
largest element available as the behavior may become glitchy and hard to
replicate once jQuery is gone.

As a drive-by fix, the iframes have a white background in case the pages
have no dedicated CSS to avoid overlapping the content of the iframes in
the "Slider" preview mode.

Resolves: #98420
Releases: main
Change-Id: I344ee7c99888261b19b97ad98cc31c519d1e1d4b
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/75850
Tested-by: core-ci <typo3@b13.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Frank Nägler <frank.naegler@typo3.com>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Frank Nägler <frank.naegler@typo3.com>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>