[TASK] Move root css variables to dedicated file
The inclusion of the global CSS variables into the
SCSS variables were flawed from the very beginning.
The include resulted in multiple unwanted occurrences
of these variables produced by the current architecture
of the SCSS files used in the backend.

It has been a known issue and accepted issue from the
start, since we did not have a plan yet to handle them.

The CSS variables are moved to a dedicated root file
that is part of the minimal setup. The minimal setup
is meant to be standalone and only included once
per page.

Resolves: #98460
Releases: main
Change-Id: I47a1cca6fb74f039e9e8df9ef6b5076f22922dc4
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/75885
Tested-by: core-ci <typo3@b13.com>
Tested-by: Markus Klein <markus.klein@typo3.org>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Markus Klein <markus.klein@typo3.org>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Benni Mack <benni@typo3.org>