[TASK] Align UI of the DB check module
DB Check seems to be the only module in backend
that uses an Overview view for displaying the functions.
Other modules have a default function that is loaded.

Replace overview page by loading Manage Reference Index as default function,
as this is probably where the least loading overhead occurs.

* Remove DatabaseIntegrityController->overviewAction
* Remove IntegrityOverview.html
* Remove Test seeOverview

Resolves: #98477
Releases: main
Change-Id: Id37d5bd50071f9abbb6bba37e6edc43f5b0a0ef2
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/75921
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Annett Jähnichen <mcmietz@web.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Annett Jähnichen <mcmietz@web.de>