[BUGFIX] Build LanguageService in TranslateViewHelper
Since #98377, the request passed to Fluid's rendering context is no
`ExtbaseRequestInterface` anymore by default, which has an impact on the
`f:translate` view helper. The view helper checks whether the request is
a ExtbaseRequestInterface and falls back to Extbase's internal language
setup.

This patch extends the view helper's `getLanguageService()` method to
check whether the `LanguageService` is already built. If not, a
`LanguageService` is built depending on the application context.

Resolves: #98501
Related: #98377
Releases: main
Change-Id: I07a831c2cb81decff207717ce2db1104581be21c
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/75941
Tested-by: core-ci <typo3@b13.com>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>