[TASK] Improve type annotations for ExpressionBuilder::literal()
PHPStan will (correctly) unify the previous type
`\PDO::PARAM_*|Connection::PARAM_*|int` to `int` (because
`\PDO::PARAM_*` as well as `Connection::PARAM_*` are all
ints).

So we're using the stricter `Connection::PARAM_*` now, which is a
superset of `\PDO::PARAM_*`.

This allows PHPStan to catch calls to this method with
incorrect values.

Resolves: #98512
Related: #98500
Releases: main, 11.5
Change-Id: I78efb486fe60aea2640435c8c48bd08ae8b3c3cd
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/76167
Tested-by: core-ci <typo3@b13.com>
Tested-by: Oliver Klee <typo3-coding@oliverklee.de>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Oliver Klee <typo3-coding@oliverklee.de>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>