[FEATURE] Introduce PSR-14 event to modify form data for edit file form
This introduces a new PSR-14 event, enabling
extension authors to modify the form data, used
by FormEngine to generate the edit file form in
the filelist module.

TYPO3 internally makes use of this event to
initialize t3editor with the suitable format options
for the file to be edited. This functionality was
previously done in a hook (#98494), which
however did no longer work due to the removal
of said hook in #97452.

Effectively this means, the new Event is an
improved replacement for the removed hook,
since the Event now provides the whole form
data array as well as the resolved FileInterface
and the current PSR-7 Request.

Resolves: #98521
Related: #98494
Related: #97452
Releases: main
Change-Id: I5de9f8ea72fcb4296f6539449f991347cbef17b6
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/75982
Tested-by: core-ci <typo3@b13.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>