[BUGFIX] Determine MariaDB Server Version correctly when prefixed
Some builds of MariaDB server and/or report server version of
MariaDB with a '5.5.5-' prefix. This change address changes
a regex to deal with this detail in the environment check

Prefixed server version information looks like:

  "MySQL 5.5.5-10.3.32-MariaDB-1:10.3.32+maria~focal-log"

compared to other MySQL servers, which don't add the '5.5.5-' prefix.

Invalid version detection will now display a dedicated message to
express that transparently.

Resolves: #98518
Releases: main
Change-Id: I60c118e9779265061fd4fc973ab21a417bdfca01
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/75991
Tested-by: core-ci <typo3@b13.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Stefan Froemken <froemken@gmail.com>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Stefan Froemken <froemken@gmail.com>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>