[FEATURE] Add TCA field control „passwordGenerator“
This patch adds a new TCA field control "passwordGenerator" for
TCA type "password", which can be used to generate a random
password, based on defined password rules.

The field control will be used in EXT:reactions the first time.

In the future the field control might be combined with the
existing password policies functionality. Therefore, most
of the new functions and classes are marked as internal
for now.

Another feature to be added in the future will be (according
to #69190) a modal / wizard, which will allow to change rules
for the password to be generated.

Resolves: #98540
Related: #98373
Releases: main
Change-Id: Iab21b8bd0b3c261a702234579f62ba841c54ea0d
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/76002
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Oliver Hader <oliver.hader@typo3.org>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>