[BUGFIX] Ensure proper error_reporting reset in JwtTrait workaround
With the introduction of CSRF-like request tokens in #97305 a
3rd party dependency has been added with a PHP8.2 deprecation.
To mitigate this, a workaround has been added to set and reset
the error level in the implementation in JwtTrait.

However, if the called methods throw exception on invalid tokens
the reset has been left out, leaving the `error_reporting` in a
changed state and thus changing the system.

This can be verified using:

> Build/Scripts/runTests.sh -s unitRandom -p 8.2 -o 1665191152

This change wraps the method calls in try-catch blocks and
ensure the error_reporting is properly reset before re-throw
the exception to keep the exception handling workflow intact.

Resolves: #98551
Releases: main
Change-Id: I692188ee9be2583bdf26f39b646784580cd783e7
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/76006
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benjamin Franzke <bfr@qbus.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Oliver Klee <typo3-coding@oliverklee.de>
Reviewed-by: Benjamin Franzke <bfr@qbus.de>