[TASK] Merge AbstractConfigurationManager away
The extbase ConfigurationManager is a complex beast full
of anti patterns. We need to refactor here to unblock
other things.

AbstractConfigurationManager is the base of the 'concrete' classes
BackendConfigurationManager and FrontendConfigurationManager. These
are @internal with ConfigurationManager being the outer API.

To reduce complexity and to allow looking at Frontend and Backend
scenarios without triggering side effects on the other one, this
divide and conquer patch simply copies all code from
AbstractConfigurationManager into the inheriting classes, making
them standalone. Existing unit tests of AbstractConfigurationManager
are also copied to the two consuming classes. No further other
refactorings are done with this patch to keep it simple.

Resolves: #98578
Releases: main
Change-Id: Ie7d663cc94c40363f46f41a09ee689113835133e
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/76089
Tested-by: core-ci <typo3@b13.com>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>