[BUGFIX] Assure a string is passed to `octdec` function
Since strict types were enabled for the GeneralUtility class, the
`GU::createDirectoryPath()` may throw an exception if
`$GLOBALS['TYPO3_CONF_VARS']['SYS']['folderCreateMask']` is not
set. This is due to the fact that the fallback value is an integer,
whereas the `octdec` function requires a string to be passed.

Resolves: #98602
Releases: main, 11.5
Change-Id: I710d28166e3d3d8e76e6125db53ea59e1b4ab26d
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/76107
Reviewed-by: Oliver Klee <typo3-coding@oliverklee.de>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Benni Mack <benni@typo3.org>