[TASK] Add .gitignore for JetBrains Fleet editor
JetBrains created a new product named `Fleet` as
lightweight editor. This tool tends to write it's
config files to `.fleet/` folders, like all of the
IntellJ based IDE's like PHPStorm uses the `.idea`
folder.

This change adds this config folder to .gitignore
to avoid adding this folder to a patch if `Fleet`
is used to create a patch.

Resolves: #98615
Releases: main, 11.5, 10.4
Change-Id: I4f629a37904c2e9fda57df830377c98a88bec69c
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/76067
Tested-by: core-ci <typo3@b13.com>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>