[TASK] Refactor Backend Live Search functionality
The backend search (aka Live Search) can search for database
records only, currently.

To make the search able to find anything other than database records a
`SearchProviderInterface` is introduced which allows to add new providers
in the future and may be used by extension authors in the long run.

As the first implementations, the current database record search is
extracted into a `DatabaseRecordProvider`, which excludes records from
the `pages` table. Additionally, a new `PageRecordProvider` is added
which comes with a custom web component that renders additional
information for pages only.

The changes introduced in this patch are not final and may change in
further iterations, therefore all interfaces are declared as "internal"
and documentation is missing on purpose.

Resolves: #92008
Releases: main
Change-Id: I16327f32a5e2bb05cc5f8a80540a2ced94213b39
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/76118
Tested-by: core-ci <typo3@b13.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Benni Mack <benni@typo3.org>