[TASK] Streamline workspaces menu and backend header styling
The active state in the workspaces menu is now streamlined and aligned
to the look the "Help" menu by removing the "tick" icon from the active
workspace and applying a blue-ish background instead.

The workspace indicator identified by the green background color is now
applied on compile time of the Fluid template. The toolbar item template
is adjusted to optionally support a `name` attribute, used to render the
workspace title compile time.

Resolves: #98636
Releases: main
Change-Id: Ifbb9cb016a331c8ece43d2038412883bf5cb3aac
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/76129
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>