[BUGFIX] Add basic validation for email and phone links
To avoid setting potentially invalid links, the "email address" and
"telephone number" fields both are marked as "required". The input type
of the former one is also changed to `email`.

As a drive-by fix, `id` and `for` properties are set to link the input
fields to their respective label.

Resolves: #98388
Releases: main, 11.5
Change-Id: I24d62f70743e2bce73dd1de3cd9ecdac5dd0343a
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/76216
Reviewed-by: Nikita Hovratov <nikita.h@live.de>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Oliver Klee <typo3-coding@oliverklee.de>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>