[BUGFIX] Apply overflow to body in iframe loaded in modal
In TYPO3 v8, a "rebounce" protection was added to deny overscrolling
the document on iOS devices (aka "rubber band effect"). This caused an
issue in #98348 where the shortcut menu opened to the wrong side and
the rubber band fix got partially reverted, breaking modals where the
content is intended to be scrollable.

Also, the scaffold is "broken" on Apple devices as `100vh` is treated as
"100vh if the address bar is shrunk", leading to a larger and thus
scrollable scaffold when using `overscroll-behavior: none` [1], which
would've been the preferred solution to solve the scrolling issue.

The solution for the scrollable scaffold are dynamic viewport
units [2], but those are not supported in all browsers, yet, and also
revert `overscroll-behavior` (tested with iPadOS 16.1).

To have the original issue fixed somehow, a workaround now adds the CSS
class `with-overflow` to the `body` tag to all iframes rendered in
modals.

[1] https://caniuse.com/css-overscroll-behavior
[2] https://caniuse.com/viewport-unit-variants

Resolves: #98710
Related: #98348
Releases: main
Change-Id: I13159cb1894aa76f81a91b8b80863a9dcfbc1970
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/76236
Tested-by: core-ci <typo3@b13.com>
Tested-by: Frank Nägler <frank.naegler@typo3.com>
Reviewed-by: Frank Nägler <frank.naegler@typo3.com>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>