[TASK] GifBuilder: Use CharRangeMap from configuration
The global charRangeMap option in TypoScript / GIFBUILDER
allows to define the pixelSpace (with a multiplicator for the fontSize)
for certain fonts.

However, it is not possible to define this per GIFBUILDER
rendering due to the access on TSFE.

Instead, the local charRangeMap of the GIFBUILDER configuration
array is used to allow the charRangeMap per generated image.

Resolves: #98720
Releases: main
Change-Id: I030ec0f4aa06190bac607bc8e8f94a0ba425d743
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/76242
Tested-by: core-ci <typo3@b13.com>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>