[BUGFIX] CKEditor 4 Anchor Button is incorrectly handled as link in CKEditor 5 migration
CKEditor 5 changed the way it handles links, and
introduced new modules to do so. In the automatic
migration from CKEditor 4 to CKEditor 5 configuration
the previously available separate button "Anchor" is
is incorrectly migrated to the new "link" module.

In CKEditor 4 there was a differentiation between
outgoing links (https://, relative etc.) and anchors on
the current page (aka #). CKEditor did not make
this differentiation anymore and introduced a unified
link handling. There is no replacement for "Anchor"
in the new configuration.

Keeping the migration would resolve in unintentionally
removing the capability of setting links in general, since
it would remove the new overarching plugin.

In addition, we are now also cleaning up the legacy
configuration after migration for removeButtons
and format_tags.

Resolves: #98714
Releases: main
Change-Id: I9f94e831135bf099dec716558120f727301a085c
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/76241
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>