[BUGFIX] Correct registration of toolbar link buttons in CKEditor 5
The CKEditor 5 has very limited support for managing links.
By default, you only get a simple input field for a string.
To bring the known and loved easy-to-use LinkBrowser to
CKEditor 5 we need to overwrite the existing linking
capabilities.

For developers and integrators, this overwrite should be
invisible as it was before with the CKEditor 4 integration.
Instead of teaching the users that they need to load custom
modules to change the link behavior, we are already migrating
this in the background for them. A typical developer can just
follow the CKEditor 5 guide to configuring the editor to its
needs and the Core is already doing the adjustments and
replacements of modules for them.

We currently overwrite the existing link and unlink commands,
but we register the toolbar buttons with a prefix. This makes
existing CKEditor configurations incompatible since the
linking buttons are missing in the toolbar.

To make migrations easier we adjust custom toolbar buttons to
match the CKEditor 5 defaults and drop the custom names for
the buttons.

In addition, we are streamlining the behavior of removing plugins.
Since we are already overwriting the "Link" plugin there is no
need to remove it if no removePlugin config is passed in the first
place. This allows us to better test the behavior and reduces the
cases we need to cover in the first place.

Resolves: #98925
Releases: main
Change-Id: I3d6d158f5ab1b21e73f86471344107a07d71b03e
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/76267
Tested-by: core-ci <typo3@b13.com>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Benni Mack <benni@typo3.org>