[TASK] Drop ineffective TMENU caches
The "hash" cache for (TMENU) menu generation is
very ineffective: The cache identifier hash contains
the page rootline, so it is single page specific.

This leads to quite some pollution of the "hash"
cache, often many more entries than page cache
entries, without high re-use / cache-hit rate.

Looking closer at this, we found the menu
generation cache to be so ineffective that it's
better to not cache at all. We first thought a
runtime cache would be a good solution but even
that one is not helpful and there are further
runtime caches on lower levels already, for
instance to resolve shortcut pages.

The patch removes this cache layer, leading to
a drastically smaller "hash" cache.

Resolves: #98964
Releases: main
Change-Id: Iadd3ad77f4d2d4683b91ca7b1f27a779ff899d26
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/76384
Tested-by: Loek Hilgersom <loek@netcoop.nl>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Loek Hilgersom <loek@netcoop.nl>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Benni Mack <benni@typo3.org>