[BUGFIX] Keep existing baseVariants for site languages
The backend UI does not support handling of baseVariants
for languages, however it is possible to define them
in the config.yaml of a site configuration.

However, when editing and saving a site, the baseVariants
manually defined were removed. This has now been changed,
as the baseVariants are kept when persisting
the site configuration via the UI.

Resolves: #93719
Releases: main, 11.5
Change-Id: I8a0b3be72ddfab46da59b3f8a69ad16121c7246b
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/76496
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Benni Mack <benni@typo3.org>