[BUGFIX] Get the correct next stage for moved elements in workspace
The method TYPO3\CMS\Workspaces\Service\StageService->getNextStageForElementCollection relies on the field t3ver_stage to calculate the next stage for changed elements in a workspace. Without it, it falls back to 0 (zero), meaning the current stage is "Editing". This fix provides the neccessary field so the next stage can be calculated correct.

Resolves: #98916
Releases: main, 11.5
Change-Id: I3b87c54b24c3cfe5eb2417fc582f3c4a3fb18ffc
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/76525
Tested-by: Josua Vogel <j.vogel97@web.de>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Josua Vogel <j.vogel97@web.de>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>