[FEATURE] Add PSR-14 event to ignore tables in Live Search
The TYPO3 backend search (aka "Live Search") has a
`\TYPO3\CMS\Backend\Search\LiveSearch\DatabaseRecordProvider` that
searches all records in database tables that have `searchFields`
configured.

In some individual cases it may not be desired to search in a certain
table, therefore a new event `BeforeSearchInDatabaseRecordProviderEvent`
is introduced to add tables to a deny list.

The event is directly used by TYPO3 to exclude the "pages" table.

Resolves: #99033
Releases: main
Change-Id: I956085d02576dae60ae6ffb011c9c3a695e08d4f
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/76522
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Oliver Bartsch <bo@cedev.de>