[TASK] Deprecate TypoScript setup "constants" TLO
This is not about TypoScript constants/settings
that are accessed using {$myConstant} in TypoScript
setup.

This is about the TypoScript setup "constants"
top-level-object: It is used in TypoScript
"parseFunc" and allows referencing string literals
by name in "parseFunc" strings, which are substituted
by using ###theName###, when property "constants = 1" is
set in "parseFunc".

This feature dates back to Kasper's "Initial revision"
and is obsolete ever since: The feature exists next
to well-known TypoScript constants, existing
usages can be easily substituted. The confusion with
TypoScript "constants" alone justifies a deprecation
of this detail.

The patch deprecates both the TypoScript setup
"constants" top-level-object, plus the "constants"
property in "parseFunc".

Resolves: #99040
Releases: main
Change-Id: I91546067db8a7eb72661087780e39f320db2cbeb
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/76532
Tested-by: core-ci <typo3@b13.com>
Tested-by: Susanne Moog <look@susi.dev>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Susanne Moog <look@susi.dev>
Reviewed-by: Benni Mack <benni@typo3.org>