[TASK] Set "keepFilename" to true by default
One of the main use cases of replacing files is updating the content of a
referenced file with new revisions or changed content. The local filename
of the replacing file in most cases is not as relevant so by default
opting to keeping the filename makes sense for the end user.

Resolves: #96442
Releases: main
Change-Id: I9fbdc7229884b48433447eb1b9616623c3a09e33
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/76533
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Oliver Bartsch <bo@cedev.de>